+++ PlatformGame/src/Render.cpp
@@ -113,7 +113,7 @@ void Render::ResetViewPort()
 }
 
 // Blit to screen
-bool Render::DrawTexture(SDL_Texture* texture, int x, int y, const SDL_Rect* section, float speed, double angle, int pivotX, int pivotY, int scaleFactor) const
+bool Render::DrawTexture(SDL_Texture* texture, int x, int y, const SDL_Rect* section, float speed, double angle, int pivotX, int pivotY, float scaleFactor) const
 {
 	bool ret = true;
 	int scale = Engine::GetInstance().window.get()->GetScale();
